Simplify OFFSET implementation
This change is based on RowNumberNode's property of keeping order of its input.
In OFFSET implementation, a RowNumberNode and a FilterNode are used for rows selection.
Before this change, if there was ORDER BY in the query, it was handled by one of
the specialised rules which ensured that sorting was replicated over RowNumberNode.
Since RowNumberNode respects sorted order of its input, it is unnecessary to add
a SortNode.
