+++ presto-main/src/main/java/io/prestosql/sql/planner/iterative/rule/ImplementOffset.java
@@ -14,12 +14,8 @@
 package io.prestosql.sql.planner.iterative.rule;
 
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
-import io.prestosql.matching.Capture;
 import io.prestosql.matching.Captures;
 import io.prestosql.matching.Pattern;
-import io.prestosql.spi.block.SortOrder;
-import io.prestosql.sql.planner.OrderingScheme;
 import io.prestosql.sql.planner.Symbol;
 import io.prestosql.sql.planner.iterative.Rule;
 import io.prestosql.sql.planner.plan.Assignments;
@@ -27,42 +23,35 @@ import io.prestosql.sql.planner.plan.FilterNode;
 import io.prestosql.sql.planner.plan.OffsetNode;
 import io.prestosql.sql.planner.plan.ProjectNode;
 import io.prestosql.sql.planner.plan.RowNumberNode;
-import io.prestosql.sql.planner.plan.SortNode;
-import io.prestosql.sql.planner.plan.TopNNode;
 import io.prestosql.sql.tree.ComparisonExpression;
 import io.prestosql.sql.tree.GenericLiteral;
 
 import java.util.Optional;
 
-import static io.prestosql.matching.Capture.newCapture;
 import static io.prestosql.spi.type.BigintType.BIGINT;
 import static io.prestosql.sql.planner.plan.Patterns.offset;
-import static io.prestosql.sql.planner.plan.Patterns.source;
-import static io.prestosql.sql.planner.plan.Patterns.topN;
 
 /**
  * Transforms:
  * <pre>
  * - Offset (row count = x)
- *    - TopN (order by a, b, c)
+ *    - Source
  * </pre>
  * Into:
  * <pre>
  * - Project (prune rowNumber symbol)
- *    - Sort (order by rowNumber)
- *       - Filter (rowNumber > x)
- *          - RowNumber
- *             - TopN (order by a, b, c)
+ *    - Filter (rowNumber > x)
+ *       - RowNumber
+ *          - Source
  * </pre>
+ * Relies on RowNumberNode's property of keeping order of its input.
+ * If the query contains an ORDER BY clause, the sorted order
+ * will be respected when leading rows are removed.
  */
-public class ImplementOffsetOverTopN
+public class ImplementOffset
         implements Rule<OffsetNode>
 {
-    private static final Capture<TopNNode> TOPN = newCapture();
-
-    private static final Pattern<OffsetNode> PATTERN = offset()
-            .with(source().matching(
-                    topN().capturedAs(TOPN)));
+    private static final Pattern<OffsetNode> PATTERN = offset();
 
     @Override
     public Pattern<OffsetNode> getPattern()
@@ -73,13 +62,11 @@ public class ImplementOffsetOverTopN
     @Override
     public Result apply(OffsetNode parent, Captures captures, Context context)
     {
-        TopNNode topN = captures.get(TOPN);
-
         Symbol rowNumberSymbol = context.getSymbolAllocator().newSymbol("row_number", BIGINT);
 
         RowNumberNode rowNumberNode = new RowNumberNode(
                 context.getIdAllocator().getNextId(),
-                topN,
+                parent.getSource(),
                 ImmutableList.of(),
                 rowNumberSymbol,
                 Optional.empty(),
@@ -93,15 +80,10 @@ public class ImplementOffsetOverTopN
                         rowNumberSymbol.toSymbolReference(),
                         new GenericLiteral("BIGINT", Long.toString(parent.getCount()))));
 
-        SortNode sortNode = new SortNode(
-                context.getIdAllocator().getNextId(),
-                filterNode,
-                new OrderingScheme(ImmutableList.of(rowNumberSymbol), ImmutableMap.of(rowNumberSymbol, SortOrder.ASC_NULLS_FIRST)));
-
         ProjectNode projectNode = new ProjectNode(
                 context.getIdAllocator().getNextId(),
-                sortNode,
-                Assignments.identity(topN.getOutputSymbols()));
+                filterNode,
+                Assignments.identity(parent.getOutputSymbols()));
 
         return Result.ofPlanNode(projectNode);
     }
